Fix controller dependency & other weblinks endpoints
