+++ tests/cypress/integration/api/com_weblinks/RateLimiting.cy.js
